pyproject.toml
[tool.poetry]
name = "telewrap"
version = "0.1.0"
description = "A Telegram bot wrapper"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "telewrap"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.2.1"
python-telegram-bot = "^21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
